Fix Tests on Master
Things got a little bit out of sync after #55 landed on master and we lost some of the safe-guards that were introduced in #64.

This PR refactors the layout of the tests, making a split between integration and unit tests with an npm script for each. It also re-instates the 'check-mocha-report' check to ensure that at least 1 test passed.