Add option to general Service Clients or just Stubs.
Allow consumers to choose if they wish to generate full Service Clients (for use with grpc-web) or just the minimal boiler plate (stubs) required when invoking grpc-web directly.

The primary driver here is to reduce overall bundle-size. Tree shaking is not an option as the generated code use CommonJS modules, not ES6 ones.